refactor: Don't change M
